Fix compilation error & add defaults
- fixes rogue compilation error in main.go
- splits types into types.go within main package
- adds defaults for max reconnect and reconnect delay for queue-
worker (subscriber)
