+++ readconfig.go
@@ -11,6 +11,10 @@ import (
 type ReadConfig struct {
 }
 
+const DefaultMaxReconnect = 120
+
+const DefaultReconnectDelay = time.Second * 2
+
 func (ReadConfig) Read() QueueWorkerConfig {
 	cfg := QueueWorkerConfig{
 		AckWait:     time.Second * 30,
@@ -58,6 +62,8 @@ func (ReadConfig) Read() QueueWorkerConfig {
 		}
 	}
 
+	cfg.MaxReconnect = DefaultMaxReconnect
+
 	if value, exists := os.LookupEnv("faas_max_reconnect"); exists {
 		val, err := strconv.Atoi(value)
 
@@ -68,11 +74,14 @@ func (ReadConfig) Read() QueueWorkerConfig {
 		}
 	}
 
+	cfg.ReconnectDelay = DefaultReconnectDelay
+
 	if value, exists := os.LookupEnv("faas_reconnect_delay"); exists {
 		reconnectDelayVal, durationErr := time.ParseDuration(value)
 
 		if durationErr != nil {
 			log.Println("parse env var: faas_reconnect_delay as time.Duration error:", durationErr)
+
 		} else {
 			cfg.ReconnectDelay = reconnectDelayVal
 		}
